requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/base.in
#
anyascii==0.3.0
    # via wagtail
asgiref==3.4.1
    # via django
beautifulsoup4==4.9.3
    # via wagtail
certifi==2021.10.8
    # via requests
charset-normalizer==2.0.10
    # via requests
django==3.2.11
    # via
    #   -r requirements/base.in
    #   django-filter
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-environ==0.8.1
    # via -r requirements/base.in
django-filter==21.1
    # via wagtail
django-modelcluster==5.2
    # via wagtail
django-taggit==1.5.1
    # via wagtail
django-treebeard==4.5.1
    # via wagtail
djangorestframework==3.13.1
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
gunicorn==20.1.0
    # via -r requirements/base.in
html5lib==1.1
    # via wagtail
idna==3.3
    # via requests
l18n==2021.3
    # via wagtail
openpyxl==3.0.9
    # via tablib
pillow==8.4.0
    # via wagtail
pytz==2021.3
    # via
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   l18n
requests==2.27.1
    # via wagtail
six==1.16.0
    # via
    #   html5lib
    #   l18n
soupsieve==2.3.1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tablib[xls,xlsx]==3.1.0
    # via wagtail
telepath==0.2
    # via wagtail
urllib3==1.26.8
    # via requests
wagtail==2.15.1
    # via -r requirements/base.in
webencodings==0.5.1
    # via html5lib
willow==1.4
    # via wagtail
xlrd==2.0.1
    # via tablib
xlsxwriter==3.0.2
    # via wagtail
xlwt==1.3.0
    # via tablib

# The following packages are considered to be unsafe in a requirements file:
# setuptools
